+++ CHANGELOG.md
@@ -15,7 +15,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Pageviews graph in article details page loads data from Journal snapshots instead of directly quering Journal API by default. Option `PAGEVIEWS_DATA_SOURCE` added to `.env` file to allow switching to old behaviour. remp/remp#442
 - `rempConfig.tracker` configuration option `explicit_referer_medium` is deprecated and renamed to `refererMedium`. The old one is still accepted due to compatibility reasons.
 - Added support for referer medium renaming - one can specify label for each medium by adding a record to `referer_medium_labels` table. remp/remp#543
-- Updated layout footer with link to REMP website. remp/remp#522
+- Updated layout footer with link to REMP website. remp/remp#522 HEAD
+- Added ability to display external events in article detail and dashboard. remp/remp#574
+- Added (optional) `article_id` field to event parameters pushed to tracker. remp/remp#556
 - Fixed property naming validator issue checking unique constraint against account names, not property names.
 - Added article conversion/pageviews stats filtering based on selected property. GH-50